backend/pyproject.toml
[tool.poetry]
name = "miami"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = "^2.8.0"
flask = "^3.0.0"
datetime = "^5.2"
flask-sqlalchemy = "^3.1.1"
get-docker-secret = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
pytest-flask = "^1.3.0"
waitress = "^2.1.2"
flask-cors = "^4.0.0"
cors = "^1.0.1"
freezegun = "^1.2.2"
mp = "^0.5.0"
ccxt = "^4.1.47"
ccxtpro = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
